feat(storybook): Migration
- Migrated UI stories to storybook project
