+++ package.json
@@ -30,7 +30,7 @@
     "build:sb": "npx lerna run build --scope @hisptz/storybook",
     "start:docs": "yarn docs start",
     "start:sb": "yarn sb start",
-    "start:sb-proxy": "yarn sb start:proxy",
+    "start:sb-proxy": "npx lerna run start:proxy --scope @hisptz/storybook",
     "test": "yarn test:lib && yarn test:component",
     "build": "npx lerna run build",
     "prepublish": "npx lerna run build --no-private",
