+++ packages/storybook/src/ui/dialogs/ConfirmDialog.stories.tsx
@@ -1,0 +1,94 @@
+import type {Story} from "@storybook/react";
+import React from "react";
+import {ConfirmDialog, ConfirmDialogProps} from "@hisptz/dhis2-ui";
+
+const Template: Story<ConfirmDialogProps> = (args) => <ConfirmDialog {...args} />;
+
+export const ConfirmDialogExample = Template.bind({});
+ConfirmDialogExample.args = {
+    title: "Confirm Title",
+    message: "Confirm Content",
+    onConfirm: () => {
+        alert("Confirm 🤗");
+    },
+    onCancel: () => {
+        alert("Cancel 😔");
+    },
+};
+
+export const ConfirmDialogExampleWithAllOptions = Template.bind({});
+ConfirmDialogExampleWithAllOptions.args = {
+    title: "Confirm Title",
+    message: "Confirm Content",
+    position: "middle",
+    size: "small",
+    hide: false,
+    cancelButtonText: "Cancel",
+    confirmButtonText: "Confirm",
+    confirmButtonColor: "primary",
+    onConfirm: () => {
+        alert("Confirm 🤗");
+    },
+    onCancel: () => {
+        alert("Cancel 😔");
+    },
+    customActions: [
+        {
+            label: "Custom Action",
+            color: "secondary",
+            onClick: () => {
+                alert("Custom Action 🙃");
+            },
+        },
+    ],
+};
+
+export default {
+    title: "Dialogs/Confirm Dialog/Confirm Dialog",
+    component: ConfirmDialogExample,
+    argTypes: {
+        title: {
+            control: {
+                type: "text",
+            },
+        },
+        message: {
+            control: {
+                type: "text",
+            },
+        },
+        position: {
+            control: {
+                type: "select",
+                options: ["top", "middle", "bottom"],
+            },
+        },
+        size: {
+            control: {
+                type: "select",
+                options: ["small", "large"],
+            },
+        },
+        hide: {
+            control: {
+                type: "boolean",
+            },
+        },
+        cancelButtonText: {
+            control: {
+                type: "text",
+            },
+        },
+        confirmButtonText: {
+            control: {
+                type: "text",
+            },
+        },
+        confirmButtonColor: {
+            control: {
+                type: "select",
+                options: ["primary", "secondary", "destructive"],
+            },
+        },
+    },
+};
