+++ packages/ui/src/components/selectors/Modals/index.tsx
@@ -1,4 +1,3 @@
-import OrgUnitSelectorModal from "./components/OrgUnitSelectorModal";
-import PeriodSelectorModal from "./components/PeriodSelectorModal";
-
-export { PeriodSelectorModal, OrgUnitSelectorModal };
+export * from "./components/PeriodSelectorModal"
+export * from "./components/OrgUnitSelectorModal"
+export * from "./types"
